# server/swagger/openapi.yaml
...
openapi: 3.0.0
servers:
- url: http://localhost:5000
info:
  description: 'This is a sample E-Commerce 2.0 app. This project requires that you build a fully-functioning e-commerce application REST API that allows users to perform various CRUD operations such as registering an account, browsing products for sale, etc.'
  version: 2.0.0
  title: E-Commerce 2.0 REST API application
  termsOfService: https://www.codecademy.com/about
  contact:
    email: easoftware@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
  "/api/users":
    post:
      tags:
      - users
      summary: Create user
      description: This is only for testing. To create a user, post to /api/auth/register.
      operationId: createUser
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '400':
          description: email already used, required value missing
        '404':
          description: user not inserted
      requestBody:
        "$ref": "#/components/requestBodies/UsersPostPut"
    get:
      tags:
      - users
      summary: Gets all users
      description: ''
      operationId: getAllUsers
      responses:
        '200':
          description: successful opertion
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Users"  
        '404':
          description: 'no user rows found'
...
  "/auth/register":
    post:
      tags:
      - auth
      summary: registers user
      description: ''
      operationId: registerUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '409': 
          description: email already in use
        '500': 
          description: could not create user
      requestBody:
        "$ref": "#/components/requestBodies/AuthRegister"
  "/auth/login":
    post:
      tags:
      - auth
      summary: logs in user
      description: Uses passort and passport-local strategy to log in user
      operationId: loginUser
      responses:
        '200':
          description: logged in user - user's id returned as a string
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
      requestBody:
        "$ref": "#/components/requestBodies/AuthLogin"
  "/auth/logout":
    get:
      tags:
      - auth
      summary: logs out user
      description: 'Uses passort and passport-local strategy to log out user'
      operationId: logoutUser
      responses:
        '205':
          description: successful opertion
        '404':
          description: 'error'
  "/auth/is_logged_in":
    get:
      tags:
      - auth
      summary: checks if user is logged in
      description: 'checks server to see if user is logged in'
      operationId: isLoggedIn
      responses:
        '200':
          description: successful opertion
        '404':
          description: 'error, suer not logged in'
  "/auth/user":
    get:
      tags:
      - auth
      summary: gets logged in user
      description: 'gets logged in user using passort and passport-local strategy'
      operationId: getLoggedInUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '404':
          description: 'no user logged in'
  "/auth/google":
    get:
      tags:
      - auth
      summary: 'logs in user using Google account'
      description: 'logs in user using OAuth 2 and Google account'
      operationId: loginUserGoogle
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '404':
          description: 'error'
  "/products":
    post:
      tags:
      - products
      summary: Create product
      description: Can only be done by admin
      operationId: createProduct
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Products"
        '400': 
          description: 'name already used, model_number already used, required value missing'
        '404':
          description: Product not inserted
      requestBody:
        "$ref": "#/components/requestBodies/Products"
    get:
      tags:
      - products
      summary: Get all products
      description: Can be done by anyone, even when not logged in
      operationId: getProducts
      responses:
        '200':
          description: successful opertion
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Products"  
        '404':
          description: 'no product rows found'
...
components:
  schemas:
...
    Products:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: Decimal
        description:
          type: string
        designer:
          type: string
...
    Users:
      type: object
      properties:
        uuid:
          type: string
        email:
          type: string
          format: email
        password_hash:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        google:
          type: integer
          format: int32
      xml:
        name: User
  requestBodies:
...
    Products:
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              category:
                type: string
              price:
                type: number
                format: Decimal
              description:
                type: string
              designer:
                type: string
...
    UsersPostPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              password_hash:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone:
                type: string
              google:
                type: integer
                format: int32
      description: 'Note: password_hash is sent. This is only ised for testing'
      required: true
    AuthRegister:
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone:
                type: string
      description: 'Note: password is sent in plaintext, and password_hash is in Response'
      required: true
    AuthLogin:
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      description: 'Note: password is sent in plaintext'
      required: true